refactor: move util methods to dedicated files
